condensing add node file
